refactor: extract Participation.SessionNotes from the context facade
Third of four. SessionNotes owns SessionNote: submit, review, revise, the reads
each surface needs, and the GDPR bulk anonymisation. The facade drops from 642
to 331 lines; 482 participation tests and 1167 web tests pass with no test file
edited.

Its moduledoc records why the three authorization questions here are not folded
into one guard, because they look like a duplicate and are not:

  submit  may you act on this session, and as whom -- the same question as
          attendance, so it calls Attendance.authorize_for_record/2
  review  is this note about one of your children -- asked of Family
  revise  are you its author -- asked of the note's own provider_id

Three authorities, one guard each. Routing review or revise through
SessionAuthorization would be the bug: that authorizer answers session-role
standing, which is not what guardianship or authorship needs.

The moduledoc also records why note writes publish through Notifications without
staging to the outbox. It reads like an omission next to the attendance writes:
it isn't. These events have no registered consumer, and :event_consumers is the
staging filter, so staging them would queue delivery work for nobody.

list_approved_notes_for_children/1 is public for the roster reads still in the
facade, and says so.
